// app/_globalRedux/services/types/alerts.d.ts
export interface AlertByDayEvents {
	starttime: string;
	endtime: string;
	vehicle_no: number | string;
	exception_type: string;
	KM: string;
	duration: string;
	startlocation: string;
	startlat: number;
	startLong: number;
	endlocation: string;
	endlat: number;
	endLong: number;
	speed: number;
	journey_statusfinal: string | null;
	Halting?: string | null;
	hour?: string;
	InvoiceNo?: string;
	InvoiceDate?: string;
	remark?: string;
	id: number;
	service_id: string;
	route_name: string;
}

export interface AlertByDateLorryData {
	sys_service_id: number;
	lorry_no: number;
	harshBreak: AlertByDayEvents[];
	harshacc: AlertByDayEvents[];
	mainpower: AlertByDayEvents[];
	MainpowerConnected: AlertByDayEvents[];
	internalPower: AlertByDayEvents[];
	overspeedKMT: AlertByDayEvents[];
	freewheeling: AlertByDayEvents[];
	freewheelingWrong: AlertByDayEvents[];
	contineousDrive: AlertByDayEvents[];
	nightdrive: AlertByDayEvents[];
	highenginetemperature: AlertByDayEvents[];
	idle: AlertByDayEvents[];
	lowengineoilpressure: AlertByDayEvents[];
	overspeed: AlertByDayEvents[];
	panic: AlertByDayEventst[];
	services: AlertByDayEvents[];
	document: AlertByDayEvents[];
	transitdelay: AlertByDayEvents[];
	unlockonmove: AlertByDayEvents[];
	PoscoOverspeed: AlertByDayEvents[];
	geofence: AlertByDayEvents[];
}